IHT-4876 Create separate files for models
