+++ app/iht/models/des/ihtReturn/Trust.scala
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2017 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package iht.models.des.ihtReturn
+
+import play.api.libs.json.Json
+
+/**
+  * Created by vineet on 06/07/17.
+  */
+case class Trust(trustName: Option[String] = None,
+                 trustUtr: Option[String] = None,
+                 trustees: Option[Set[Trustee]] = None,
+                 trustAssets: Option[Set[Asset]] = None,
+                 trustLiabilities: Option[Set[Liability]] = None,
+                 trustExemptions: Option[Set[Exemption]] = None)
+
+object Trust {
+  implicit val formats = Json.format[Trust]
+}
