+++ test/iht/testhelpers/CommonBuilder.scala
@@ -16,6 +16,8 @@
 
 package iht.testhelpers
 
+import iht.models.des.ihtReturn.{Declaration, IHTReturn}
+
 import scala.util.Random
 import iht.constants.IhtProperties
 import iht.models.application.{ApplicationDetails, IhtApplication, ProbateDetails}
@@ -27,7 +29,6 @@ import iht.models.application.exemptions._
 import iht.models.application.tnrb._
 import iht.models.{ReturnDetails, _}
 import iht.utils.{CommonHelper, KickOutReason, ApplicationStatus => AppStatus}
-import models.des.iht_return.{Declaration, IHTReturn}
 import models.des.{Deceased, Event, EventRegistration}
 import org.joda.time.{DateTime, LocalDate}
 import org.mockito.invocation.InvocationOnMock
